pyproject.toml
[tool.poetry]
name = "cs224g"
version = "0.1.0"
description = ""
authors = [
    "winsonc7 <winsonc@stanford.edu>"
]

readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = ">=1.61.1,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
flask = ">=3.1.0,<4.0.0"
flask-cors = ">=4.0.0,<5.0.0"
elevenlabs = ">=1.50.7,<2.0.0"
pyaudio = ">=0.2.14,<0.3.0"
google-cloud-speech = "^2.31.0"
grpcio = "1.60.1"
pynput = "^1.7.7"
sounddevice = "^0.5.1"
google-cloud-texttospeech = "^2.25.0"
numpy = "^2.2.3"
supabase = "^2.13.0"
ics = "^0.7.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
